[refactor] UserError.java 경로 변경
error 폴더를 만들어 그곳으로 이동
