+++ backend/src/main/java/com/ddbb/dingdong/domain/user/service/error/UserErrors.java
@@ -1,4 +1,4 @@
-package com.ddbb.dingdong.domain.user.service;
+package com.ddbb.dingdong.domain.user.service.error;
 
 import com.ddbb.dingdong.domain.common.exception.ErrorInfo;
